backend/requirements.txt
asgiref==3.10.0
coverage==7.10.7
Django==4.2
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
